+++ src/FMBot.Persistence.Domain/Models/UserDiscogsReleases.cs
@@ -8,7 +8,7 @@ public class UserDiscogsReleases
 
     public int UserId { get; set; }
 
-    public int InstanceId { get; set; }
+    public long InstanceId { get; set; }
 
     public DateTime DateAdded { get; set; }
     public int? Rating { get; set; }
